LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.BlobURLDedicatedWorker.sub.https.tentative.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Unsupported permission name "storage-access".

Harness Error (TIMEOUT), message = null

TIMEOUT Verify that if the third-party context creates a blob URL using the StorageAccessHandle and sends it to the dedicated worker, the dedicated worker fetch succeeds. Test timed out
